requirements.txt
fastapi==0.68.0
pydantic==1.8.2
openai==0.10.2
pymongo==3.12.0
python-dotenv==0.19.0
uvicorn==0.15.0
